finished testing distance calc

+++ graph/edge.h
@@ -54,7 +54,7 @@ class Edge
      * @param w - the weight of the edge
      * @param lbl - the edge label
      */
-    Edge(Vertex u, Vertex v, int w, string lbl)
+    Edge(Vertex u, Vertex v, double w, string lbl)
         : source(u), dest(v), label(lbl), weight(w)
     { /* nothing */
     }
@@ -88,7 +88,7 @@ class Edge
     /**
      * Gets edge weight.
      */
-    int getWeight() const
+    double getWeight() const
     {
         return this->weight;
     }
@@ -137,7 +137,7 @@ class Edge
 
 private:
     string label; /**< The edge label **/
-    int weight; /**< The edge weight (if in a weighed graph) **/
+    double weight; /**< The edge weight (if in a weighed graph) **/
 
     //unordered_set to store airlines
     std::unordered_set<string> airlines_;
